split src
